ai-vision/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
ultralytics==8.1.36
onnxruntime==1.17.3
Pillow==10.3.0
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4
python-json-logger==2.0.7
